feat: context code clean up and adding model setter function
